+++ src/sssom_pydantic/database/__init__.py
@@ -13,10 +13,9 @@ from __future__ import annotations
 
 import contextlib
 import datetime
-from abc import ABC, abstractmethod
 from collections.abc import Callable, Collection, Generator, Iterable, Sequence
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, Concatenate, Literal, ParamSpec, cast, overload
+from typing import TYPE_CHECKING, Any, ClassVar, Literal, ParamSpec, overload
 
 import curies
 import sqlmodel
@@ -37,8 +36,9 @@ from typing_extensions import Self
 import sssom_pydantic
 from sssom_pydantic import MappingTool, Metadata, SemanticMapping
 from sssom_pydantic.api import MappingSet, MappingSetRecord, SemanticMappingHash
+from sssom_pydantic.database.repo import SemanticMappingRepository
 from sssom_pydantic.models import Cardinality
-from sssom_pydantic.process import UNSURE, Mark, curate, publish
+from sssom_pydantic.process import UNSURE
 from sssom_pydantic.query import Query
 
 if TYPE_CHECKING:
@@ -201,107 +201,6 @@ class SemanticMappingModel(SQLModel, table=True):
         return SemanticMapping.model_validate(d)
 
 
-class SemanticMappingRepository(ABC):
-    """Interact with a repository of semantic mappings."""
-
-    @abstractmethod
-    def hash_mapping(self, mapping: SemanticMapping) -> Reference:
-        """Get a reference for the mapping."""
-
-    @abstractmethod
-    def count_mappings(
-        self, where_clauses: Query | list[ColumnExpressionArgument[bool]] | None = None
-    ) -> int:
-        """Count the mappings in the database."""
-
-    def add_mapping(self, mapping: SemanticMapping) -> Reference:
-        """Add a mapping to the database."""
-        rv = self.add_mappings([mapping])
-        return rv[0]
-
-    @abstractmethod
-    def add_mappings(self, mappings: Iterable[SemanticMapping]) -> list[Reference]:
-        """Add mappings to the database."""
-
-    @abstractmethod
-    def delete_mapping(self, reference: Reference | SemanticMapping) -> None:
-        """Delete a mapping from the database."""
-
-    # docstr-coverage:excused `overload`
-    @overload
-    def get_mapping(
-        self, reference: Reference, *, strict: Literal[True] = ...
-    ) -> SemanticMappingModel: ...
-
-    # docstr-coverage:excused `overload`
-    @overload
-    def get_mapping(
-        self, reference: Reference, *, strict: Literal[False] = ...
-    ) -> SemanticMappingModel | None: ...
-
-    @abstractmethod
-    def get_mapping(
-        self, reference: Reference, *, strict: bool = False
-    ) -> SemanticMappingModel | None:
-        """Get a mapping."""
-
-    @abstractmethod
-    def get_mappings(
-        self,
-        where_clauses: Query | list[ColumnExpressionArgument[bool]] | None = None,
-        limit: int | None = None,
-        offset: int | None = None,
-        order_by: ColumnExpressionArgument[Any] | list[ColumnExpressionArgument[Any]] | None = None,
-    ) -> Sequence[SemanticMappingModel]:
-        """Get mappings."""
-
-    def curate(
-        self,
-        reference: Reference,
-        authors: Reference | list[Reference],
-        mark: Mark,
-        confidence: float | None = None,
-        add_date: bool = True,
-        **kwargs: Any,
-    ) -> Reference:
-        """Curate a mapping and return the new mapping's record."""
-        if isinstance(authors, Reference):
-            authors = [authors]
-        return self._mutate(
-            reference,
-            curate,
-            authors=authors,
-            mark=mark,
-            confidence=confidence,
-            add_date=add_date,
-            **kwargs,
-        )
-
-    def publish(
-        self,
-        reference: Reference,
-        date: datetime.date | None = None,
-    ) -> Reference:
-        """Publish a mapping and return the new mapping's record."""
-        return self._mutate(reference, publish, date=date)
-
-    def _mutate(
-        self,
-        reference: Reference,
-        f: Callable[Concatenate[SemanticMapping, P], SemanticMapping],
-        *args: P.args,
-        **kwargs: P.kwargs,
-    ) -> Reference:
-        mapping = self.get_mapping(reference)
-        if mapping is None:
-            raise KeyError
-        new_mapping = f(mapping.to_semantic_mapping(), *args, **kwargs)
-        new_mapping = new_mapping.model_copy(update={"record": self.hash_mapping(new_mapping)})
-        self.add_mapping(new_mapping)
-        self.delete_mapping(reference)
-        return cast(Reference, new_mapping.record)
-
-
 class SemanticMappingDatabase(SemanticMappingRepository):
     """A repository of semantic mappings in a SQL database, implemented using :mod:`sqlalchemy`."""
 
@@ -411,33 +310,28 @@ class SemanticMappingDatabase(SemanticMappingRepository):
                 session.delete(obj)
                 session.commit()
 
-    def _ensure(self, reference: Reference | SemanticMapping) -> Reference:
-        if isinstance(reference, SemanticMapping):
-            return self.hash_mapping(reference)
-        return reference
-
     # docstr-coverage:excused `overload`
     @overload
     def get_mapping(
         self, reference: Reference, *, strict: Literal[True] = ...
-    ) -> SemanticMappingModel: ...
+    ) -> SemanticMapping: ...
 
     # docstr-coverage:excused `overload`
     @overload
     def get_mapping(
         self, reference: Reference, *, strict: Literal[False] = ...
-    ) -> SemanticMappingModel | None: ...
+    ) -> SemanticMapping | None: ...
 
-    def get_mapping(
-        self, reference: Reference, *, strict: bool = False
-    ) -> SemanticMappingModel | None:
+    def get_mapping(self, reference: Reference, *, strict: bool = False) -> SemanticMapping | None:
         """Get a mapping."""
         with self.get_session() as session:
-            s = session.exec(self._get_mapping_by_reference(reference))
-            if strict:
-                return s.one()
+            rv = session.exec(self._get_mapping_by_reference(reference)).first()
+            if rv is not None:
+                return rv.to_semantic_mapping()
+            elif strict:
+                raise ValueError
             else:
-                return s.first()
+                return None
 
     def get_mappings(
         self,
@@ -445,7 +339,7 @@ class SemanticMappingDatabase(SemanticMappingRepository):
         limit: int | None = None,
         offset: int | None = None,
         order_by: ColumnExpressionArgument[Any] | list[ColumnExpressionArgument[Any]] | None = None,
-    ) -> Sequence[SemanticMappingModel]:
+    ) -> Sequence[SemanticMapping]:
         """Get mappings."""
         with self.get_session() as session:
             statement = select(SemanticMappingModel)
@@ -469,7 +363,7 @@ class SemanticMappingDatabase(SemanticMappingRepository):
             else:
                 statement = statement.order_by(order_by)
 
-            return session.exec(statement).all()
+            return [mapping.to_semantic_mapping() for mapping in session.exec(statement).all()]
 
     def read(
         self,
@@ -497,9 +391,8 @@ class SemanticMappingDatabase(SemanticMappingRepository):
     ) -> None:
         """Write the database to a file."""
         mappings = self.get_mappings(where_clauses=where_clauses, limit=limit, offset=offset)
-        mapping_it = (m.to_semantic_mapping() for m in mappings)
         sssom_pydantic.write(
-            mapping_it,
+            mappings,
             path,
             metadata=metadata,
             converter=converter,
